Learn more (#9)
Implementing Learn More. Added code to store facts. The facts are initially stored in a static file and are loaded into memory. Once the user leaves the app, the facts and their states are stored in a file on the device. When the user reopens the app, the facts state's are loaded.