Speed (#22)
There was a lot of code before to save a bitmap internally while the app didn't have permission to save externally, but it wasn't really necessary. I changed the code to store the rendered Bitmap directly in the submission model, in that way we are not converting between stored images and bitmap's. I pushed the permissions check up into fragment load. I pushed the saving to external storage into a separate thread. The process produced a 2x speedup (4 seconds to 2 seconds from image capture to next fragment loaded).

Also added a lot of comments.